backend/requirements.txt
fastapi
uvicorn
websockets
python-multipart
pandas
openpyxl
langchain
langgraph
langchain-openai
matplotlib
seaborn
plotly
kaleido
python-dotenv
pydantic-settings
tabulate
statsmodels
scikit-learn
